requirements.txt
asgiref==3.4.1
certifi==2021.5.30
chardet==4.0.0
Django==3.2.5
django-crispy-forms==1.12.0
django-environ==0.4.5
idna==2.10
mysqlclient==2.0.3
Pillow==8.3.1
psycopg2==2.9.1
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
stripe==2.59.0
urllib3==1.26.6
